Fix warnings with encoded_size

+++ x86_helpers.cpp
@@ -128,7 +128,7 @@ uint32_t Push(xed_state_t *dstate, xed_reg_enum_t r, unsigned char *encoded, siz
   xed_encoder_request_set_reg(&push, XED_OPERAND_REG0, GetFullSizeRegister(r, dstate->stack_addr_width));
   xed_encoder_request_set_operand_order(&push, 0, XED_OPERAND_REG0);
 
-  xed_error = xed_encode(&push, encoded, encoded_size, &olen);
+  xed_error = xed_encode(&push, encoded, (unsigned int)encoded_size, &olen);
   if (xed_error != XED_ERROR_NONE) {
     FATAL("Error encoding instruction");
   }
@@ -151,7 +151,7 @@ uint32_t Pop(xed_state_t *dstate, xed_reg_enum_t r, unsigned char *encoded, size
   xed_encoder_request_set_reg(&pop, XED_OPERAND_REG0, GetFullSizeRegister(r, dstate->stack_addr_width));
   xed_encoder_request_set_operand_order(&pop, 0, XED_OPERAND_REG0);
 
-  xed_error = xed_encode(&pop, encoded, encoded_size, &olen);
+  xed_error = xed_encode(&pop, encoded, (unsigned int)encoded_size, &olen);
   if (xed_error != XED_ERROR_NONE) {
     FATAL("Error encoding instruction");
   }
@@ -228,7 +228,7 @@ uint32_t Mov(xed_state_t *dstate, uint32_t operand_width, xed_reg_enum_t base_re
   xed_encoder_request_set_reg(&mov, XED_OPERAND_REG0, r2);
   xed_encoder_request_set_operand_order(&mov, 1, XED_OPERAND_REG0);
 
-  xed_error = xed_encode(&mov, encoded, encoded_size, &olen);
+  xed_error = xed_encode(&mov, encoded, (unsigned int)encoded_size, &olen);
   if (xed_error != XED_ERROR_NONE) {
     FATAL("Error encoding instruction");
   }
@@ -253,7 +253,7 @@ uint32_t Lzcnt(xed_state_t *dstate, uint32_t operand_width, xed_reg_enum_t dest_
   xed_encoder_request_set_reg(&lzcnt, XED_OPERAND_REG1, src_reg);
   xed_encoder_request_set_operand_order(&lzcnt, 1, XED_OPERAND_REG1);
 
-  xed_error = xed_encode(&lzcnt, encoded, encoded_size, &olen);
+  xed_error = xed_encode(&lzcnt, encoded, (unsigned int)encoded_size, &olen);
   if (xed_error != XED_ERROR_NONE) {
     FATAL("Error encoding instruction");
   }
@@ -278,7 +278,7 @@ uint32_t CmpImm8(xed_state_t *dstate, uint32_t operand_width, xed_reg_enum_t des
   xed_encoder_request_set_uimm0_bits(&lzcnt, imm, 8);
   xed_encoder_request_set_operand_order(&lzcnt, 1, XED_OPERAND_IMM0);
 
-  xed_error = xed_encode(&lzcnt, encoded, encoded_size, &olen);
+  xed_error = xed_encode(&lzcnt, encoded, (unsigned int)encoded_size, &olen);
   if (xed_error != XED_ERROR_NONE) {
     FATAL("Error encoding instruction");
   }
